requirements.txt
pyfrc
pynetworktables
robotpy-ctre
robotpy-navx
robotpy-rev
robotpy-pathfinder
